emotion-dubbing/docker/model-hub/requirements.txt
huggingface_hub
torch
transformers
numpy
pyyaml
